LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowfontfrom.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
  <body>
    <script src="../resources/securitypolicy-tests-base.js"></script>
    <script>
      if (document.SecurityPolicy.allowsFontFrom('http://example.com/'))
          log('PASS font is allowed when no policy exists.');
      else
          log('FAIL font is not allowed when no policy exists.');

      injectPolicy("font-src http://notexample.com;");

      if (!document.SecurityPolicy.allowsFontFrom('http://example.com/'))
          log('PASS font is not allowed when policy exists.');
      else
          log('PASS font is allowed when policy exists.');
    </script>
  </body>
</html>
