LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowfontfrom-expected.txt

PASS font is allowed when no policy exists.
PASS font is not allowed when policy exists.
